Got the CCC data out of the calibration-module
